Modified tasks for more streamlined queue management, also fixed some… (#10)
* Modified tasks for more streamlined queue management, also fixed some of the issues with the screen labels being too small for the inputed data.

* Addressed comments made for merge request, added doxygen style comments to newly made api

+++ CM7/UI/dashboard_ui.c
@@ -52,7 +52,7 @@ void DashboardUI_SetBrake(uint8_t brake_percent)
     );
 }
 
-void DashboardUI_SetPower(uint8_t power){
+void DashboardUI_SetPower(uint16_t power){
 	lv_label_set_text_fmt(
 		ui_PowerBaseTag,
 		"%u",
